Should be ints
Still achieving kattis score of 15. I think we should be able to use board.reverse() to also save the opposite of the state into our hashmap somehow to save time on iterations.

+++ checkers/Player.java
@@ -117,7 +117,7 @@ public class Player
         return numWhite - numRed;
     }
 
-    private StateAndScore alphaBeta( GameState state, int depth, double alpha, double beta, int player )
+    private StateAndScore alphaBeta( GameState state, int depth, int alpha, int beta, int player )
     {
         Vector<GameState> nextStates = new Vector<GameState>();
         state.findPossibleMoves( nextStates );
